backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
 
# Database
mysql-connector-python==8.4.0
 
# Auth & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==5.0.0
 
# Config
python-dotenv==1.0.1
pydantic==2.11.1
pydantic-settings==2.8.1
 
# File handling
python-multipart==0.0.9
aiofiles==23.2.1
Pillow==11.1.0
 
# HTTP client
httpx==0.27.0
requests==2.33.1
 
# Payment
razorpay==2.0.1
 
# PDF generation
reportlab==4.2.0
 
# Optional (only if you really use them)
beautifulsoup4==4.14.3
playwright==1.58.0
orjson==3.11.8
ujson==5.12.0